requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
email-validator>=2.1.1
python-dotenv>=1.0.1

sqlalchemy>=2.0.30
alembic>=1.13.1
psycopg2-binary>=2.9.9

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1

pandas
numpy
scikit-learn
yfinance
joblib
xgboost>=2.0.0
torch>=2.2.0
